refactor : DistanceFare enum 제거 후 distanceFare 전략패턴으로 통합
